Gulp composite tasks in a separate file
