+++ gulpfile.js
@@ -1,7 +1,4 @@
 var
-  path = require("path");
-
-var
   gulp = require("gulp"),
   gulpHelp = require("gulp-help");
 
@@ -14,38 +11,7 @@ var
 
 gulpHelp(gulp); // Update `gulp.task` signature to take task descriptions.
 
-
-/**
- * Composite Tasks
- */
-
-gulp.task("default", false, function () {
-  gulp.tasks.help.fn();
-});
-
-gulp.task("build", "Copy assets, build CSS and JS.", ["frontend-clean"], function () {
-  gulp.run("copy");
-  gulp.run("build:css");
-  gulp.run("build:js");
-});
-
-gulp.task("build-dev", "Build, but with unminified JS + sourcemaps.", ["frontend-clean"],
-  function () {
-    gulp.run("copy");
-    gulp.run("build:css");
-    gulp.run("build:js-dev");
-  }
-);
-
-gulp.task("watch", "Perform build-dev when sources change.", ["build-dev"], function () {
-  gulp.watch(path.join(config.frontendFullPath, "**/*"), ["build-dev"]);
-});
-
-
-/**
- * Component Tasks
- */
-
+require("./tasks/composite")(gulp, config);
 require("./tasks/lint")(gulp, config);
 require("./tasks/frontend/clean")(gulp, config);
 require("./tasks/frontend/copy")(gulp, config);
